# backend/event-service/event_openapi.yaml
openapi: 3.0.3
info:
  title: Event Service API
  description: API quản lý sự kiện và người dùng tham gia sự kiện trong hệ thống VolunteerHub
  version: 1.0.0
  contact:
    name: DongVu1105
    email: admin@gmail.com
servers:
  - url: http://localhost:8084/event
    description: Development server
...
paths:
  /create:
    post:
      tags:
        - Event Management
      summary: Tạo sự kiện mới
      description: Endpoint cho phép quản lý sự kiện tạo một sự kiện mới với thông tin chi tiết và hình ảnh (tùy chọn)
      operationId: createEvent
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - request
              properties:
                request:
                  type: string
                  description: JSON string chứa thông tin sự kiện
                  example: '{"title":"Sự kiện từ thiện","location":"Hà Nội","description":"Mô tả sự kiện","startDate":"2025-12-01T08:00:00Z","finishDate":"2025-12-02T17:00:00Z","category":"Từ thiện"}'
                file:
                  type: string
                  format: binary
                  description: File hình ảnh sự kiện (tùy chọn)
      responses:
        '200':
          description: Tạo sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
...
  /delete/{eventId}:
    delete:
      tags:
        - Event Management
      summary: Xóa sự kiện
      description: Endpoint cho phép admin xóa một sự kiện
      operationId: deleteEvent
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID của sự kiện cần xóa
          schema:
            type: string
      responses:
        '200':
          description: Xóa sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
components:
  schemas:
    ApiResponseEventResponse:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          $ref: '#/components/schemas/EventResponse'
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          default: 0
        message:
          type: string
          default: success!
        data:
          type: object
          nullable: true
...
    EventResponse:
      type: object
      properties:
        id:
          type: string
          description: ID của sự kiện
        title:
          type: string
          description: Tiêu đề sự kiện
        location:
          type: string
          description: Địa điểm tổ chức sự kiện
        description:
          type: string
          description: Mô tả chi tiết về sự kiện
        startDate:
          type: string
          format: date-time
          description: Ngày giờ bắt đầu sự kiện
        finishDate:
          type: string
          format: date-time
          description: Ngày giờ kết thúc sự kiện
        category:
          type: string
          description: Danh mục sự kiện
        managerId:
          type: string
          description: ID của người quản lý sự kiện
        managerUsername:
          type: string
          description: Tên đăng nhập của người quản lý sự kiện
        createdDate:
          type: string
          description: Ngày tạo sự kiện
        image:
          type: string
          description: URL hình ảnh sự kiện
        statusEvent:
          type: boolean
          description: Trạng thái duyệt của sự kiện
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []